Implement very basic migration
This implements very basic migration from AgileZen to Pivotal Tracker.

It's only going to grab your stories and insert them in Tracker, at this
point... it will not mark them complete or in-progress, so it's best used
for migration without having started to use AgileZen. In my case, I had
used AgileZen to organize my thoughts for an estimation, so I hadn't begun
moving stories or anything, yet.

It WILL, however copy comments & tasks, now. For comments, it will not
be able to mark them as having been from the user they're assigned to in
AgileZen, (any comment added via the API is assigned to the Tracker user
to whom the API key belongs). To alieviate this, it adds a small info
block to the beginning of each comment, showing who made it, and when.

I would recommend creating a Tracker user called "AgileZen Migrator" or
something similar, just for the migration, lest everything be assigned
to your user.